+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.1.0
+
+- Add list command
+- Add install command
+- Add uninstall command
+- Add colored UI
+
+## 0.0.2
+
+- Refactor to serve command
+- Add help texts
+- Add version
+
 ## 0.0.1
 
 - Initial commit (cloned https://github.com/robbiev/devdns)